shitcode

+++ plugins/wlx/mpv/src/wlxmpv.c
@@ -9,25 +9,19 @@
 #include <string.h>
 #include "wlxplugin.h"
 
-#define _detectstring  "(EXT=\"AVI\")|(EXT=\"MKV\")|(EXT=\"FLV\")|(EXT=\"MPG\")|\
-(EXT=\"MPEG\")|(EXT=\"MP4\")|(EXT=\"3GP\")|(EXT=\"MP3\")|(EXT=\"OGG\")|(EXT=\"WMA\")|\
-(EXT=\"BIK\")|(EXT=\"VOC\")|(EXT=\"WAV\")|(EXT=\"WEBM\")|(EXT=\"VOB\")|(EXT=\"ROQ\")|\
-(EXT=\"IVF\")|(EXT=\"MOV\")|(EXT=\"FLAC\")|(EXT=\"WMV\")"
-#define _defaultparams "--force-window=yes --keep-open=yes --script-opts=osc-visibility=always\
- --cursor-autohide-fs-only"
-#define _defaultcmd "mpv"
-#define _configfile "settings.ini"
+#define MPV_PARAMS "--force-window=yes --keep-open=yes --script-opts=osc-visibility=always --cursor-autohide-fs-only"
 
+gboolean is_pkill = FALSE;
 static char cfg_path[PATH_MAX];
 
-gchar *get_file_ext(const gchar *Filename)
+gchar *get_file_ext(const gchar *filename)
 {
-	if (g_file_test(Filename, G_FILE_TEST_IS_DIR))
+	if (g_file_test(filename, G_FILE_TEST_IS_DIR))
 		return NULL;
 
 	gchar *basename, *result, *tmpval;
 
-	basename = g_path_get_basename(Filename);
+	basename = g_path_get_basename(filename);
 	result = g_strrstr(basename, ".");
 
 	if (result)
@@ -47,31 +41,31 @@ gchar *get_file_ext(const gchar *Filename)
 	return result;
 }
 
-gchar *cfg_get_value(GKeyFile *Cfg, const gchar *Key, const gchar *FileExt, const gchar *ContenType)
+gchar *cfg_get_value(GKeyFile *cfg, const gchar *key, const gchar *ext, const gchar *content)
 {
 	gchar *result = NULL, *mediatype;
 
-	if (FileExt)
-		result = g_key_file_get_string(Cfg, FileExt, Key, NULL);
+	if (ext)
+		result = g_key_file_get_string(cfg, ext, key, NULL);
 
-	if (!result && ContenType)
+	if (!result && content)
 	{
-		result = g_key_file_get_string(Cfg, ContenType, Key, NULL);
+		result = g_key_file_get_string(cfg, content, key, NULL);
 
 		if (!result)
 		{
-			mediatype = g_strdup_printf("%.5s", ContenType);
+			mediatype = g_strdup_printf("%.5s", content);
 
 			if (mediatype)
 			{
-				result = g_key_file_get_string(Cfg, mediatype, Key, NULL);
+				result = g_key_file_get_string(cfg, content, key, NULL);
 				g_free(mediatype);
 			}
 		}
 	}
 
 	if (!result)
-		result = g_key_file_get_string(Cfg, "Default", Key, NULL);
+		result = g_key_file_get_string(cfg, "Default", key, NULL);
 
 
 	return result;
@@ -82,8 +76,9 @@ HWND DCPCALL ListLoad(HWND ParentWin, char* FileToLoad, int ShowFlags)
 	GKeyFile *cfg;
 	GError *err = NULL;
 	gboolean is_certain = FALSE;
-	gboolean bval = FALSE;
-	gchar *cmdstr, *params;
+	gboolean is_socket = FALSE;
+	gchar *exec = NULL;
+	gchar *params = NULL;
 	GdkNativeWindow id;
 	GtkWidget *gFix;
 	GtkWidget *mpv;
@@ -93,32 +88,22 @@ HWND DCPCALL ListLoad(HWND ParentWin, char* FileToLoad, int ShowFlags)
 	cfg = g_key_file_new();
 
 	if (!g_key_file_load_from_file(cfg, cfg_path, G_KEY_FILE_KEEP_COMMENTS, &err))
-	{
-		g_print("mpv.wlx (%s): %s\n", cfg_path, (err)->message);
-		cmdstr = _defaultcmd;
-		params = _defaultparams;
-	}
+		g_print("%s (%s): %s\n", PLUGNAME, cfg_path, (err)->message);
 	else
 	{
 		gchar *ext = get_file_ext(FileToLoad);
 		gchar *content_type = g_content_type_guess(FileToLoad, NULL, 0, &is_certain);
 
-		bval = g_key_file_get_boolean(cfg, "Default", "PrintContentType", NULL);
+		if (g_key_file_get_boolean(cfg, "Default", "PrintContentType", NULL))
+			g_print("%s (%s): content_type = %s\n", PLUGNAME, FileToLoad, content_type);
 
-		if (bval)
-			g_print("content_type = %s\n", content_type);
+		is_socket = g_key_file_get_boolean(cfg, "Default", "GTK_Socket", NULL);
+		is_pkill = g_key_file_get_boolean(cfg, "Default", "Use_pkill", NULL);
 
-		bval = g_key_file_get_boolean(cfg, "Default", "GTK_Socket", NULL);
-
-		cmdstr = cfg_get_value(cfg, "Cmd", ext, content_type);
-
-		if (!cmdstr)
-			cmdstr = _defaultcmd;
-
+		exec = cfg_get_value(cfg, "Cmd", ext, content_type);
 		params = cfg_get_value(cfg, "Params", ext, content_type);
-
-		if (!params)
-			params = _defaultparams;
+		g_free(content_type);
+		g_free(ext);
 	}
 
 	g_key_file_free(cfg);
@@ -130,7 +115,7 @@ HWND DCPCALL ListLoad(HWND ParentWin, char* FileToLoad, int ShowFlags)
 	gFix = gtk_vbox_new(FALSE, 5);
 	gtk_container_add(GTK_CONTAINER(GTK_WIDGET(ParentWin)), gFix);
 
-	if (bval)
+	if (is_socket)
 	{
 		mpv = gtk_socket_new();
 		gtk_container_add(GTK_CONTAINER(gFix), mpv);
@@ -148,7 +133,11 @@ HWND DCPCALL ListLoad(HWND ParentWin, char* FileToLoad, int ShowFlags)
 		gtk_widget_modify_bg(mpv, GTK_STATE_NORMAL, &color);
 	}
 
-	gchar *command = g_strdup_printf("%s %s --wid=%d %s", cmdstr, params, id, g_shell_quote(FileToLoad));
+	gchar *quoted = g_shell_quote(FileToLoad);
+	gchar *command = g_strdup_printf("%s %s --wid=%d %s", exec ? exec : "mpv", params ? params : MPV_PARAMS, id, quoted);
+	g_free(exec);
+	g_free(params);
+	g_free(quoted);
 
 	if (!g_shell_parse_argv(command, NULL, &argv, NULL))
 	{
@@ -161,9 +150,12 @@ HWND DCPCALL ListLoad(HWND ParentWin, char* FileToLoad, int ShowFlags)
 
 	if ((id == 0) || (!g_spawn_async(NULL, argv, NULL, G_SPAWN_SEARCH_PATH | G_SPAWN_DO_NOT_REAP_CHILD, NULL, NULL, &mpv_pid, NULL)))
 	{
+		g_strfreev(argv);
 		gtk_widget_destroy(gFix);
 		return NULL;
 	}
+	
+	g_strfreev(argv);
 
 	g_object_set_data(G_OBJECT(gFix), "pid", GINT_TO_POINTER(mpv_pid));
 
@@ -174,30 +166,37 @@ HWND DCPCALL ListLoad(HWND ParentWin, char* FileToLoad, int ShowFlags)
 
 void DCPCALL ListCloseWindow(HWND ListWin)
 {
-	int status;
-	pid_t mpv_pid = (pid_t)GPOINTER_TO_INT(g_object_get_data(G_OBJECT(ListWin), "pid"));
-	kill(mpv_pid, SIGTERM);
-	g_print("ListCloseWindow: mpv_pid=%d, waitpid=%d, exit_status=%d\n", mpv_pid, waitpid(mpv_pid, &status, 0), WEXITSTATUS(status));
+	if (is_pkill)
+	{
+		int status;
+		pid_t mpv_pid = (pid_t)GPOINTER_TO_INT(g_object_get_data(G_OBJECT(ListWin), "pid"));
+		kill(mpv_pid, SIGTERM);
+		g_print("ListCloseWindow: mpv_pid=%d, waitpid=%d, exit_status=%d\n", mpv_pid, waitpid(mpv_pid, &status, 0), WEXITSTATUS(status));
+	}
+
 	gtk_widget_destroy(GTK_WIDGET(ListWin));
 }
 
 void DCPCALL ListGetDetectString(char* DetectString, int maxlen)
 {
 	GKeyFile *cfg;
-	gchar *_detectstr;
+	gchar *string = NULL;
 
 	cfg = g_key_file_new();
 
 	if (!g_key_file_load_from_file(cfg, cfg_path, G_KEY_FILE_KEEP_COMMENTS, NULL))
-		g_strlcpy(DetectString, _detectstring, maxlen - 1);
+		g_strlcpy(DetectString, DETECT_STRING, maxlen - 1);
 	else
 	{
-		_detectstr = g_key_file_get_string(cfg, "Default", "DetectString", NULL);
+		string = g_key_file_get_string(cfg, "Default", "DetectString", NULL);
 
-		if (!_detectstr)
-			g_strlcpy(DetectString, _detectstring, maxlen - 1);
+		if (!string)
+			g_strlcpy(DetectString, DETECT_STRING, maxlen - 1);
 		else
-			g_strlcpy(DetectString, _detectstr, maxlen - 1);
+		{
+			g_strlcpy(DetectString, string, maxlen - 1);
+			g_free(string);
+		}
 	}
 
 	g_key_file_free(cfg);
@@ -216,7 +215,7 @@ int DCPCALL ListSendCommand(HWND ListWin, int Command, int Parameter)
 void DCPCALL ListSetDefaultParams(ListDefaultParamStruct* dps)
 {
 	Dl_info dlinfo;
-
+	gchar cfg_name[] = "settings.ini"; 
 	memset(&dlinfo, 0, sizeof(dlinfo));
 
 	if (dladdr(cfg_path, &dlinfo) != 0)
@@ -225,6 +224,6 @@ void DCPCALL ListSetDefaultParams(ListDefaultParamStruct* dps)
 		char *pos = strrchr(cfg_path, '/');
 
 		if (pos)
-			strcpy(pos + 1, _configfile);
+			strcpy(pos + 1, cfg_name);
 	}
 }
